IGNITE-13492: fix comments
