+++ modules/ducktests/tests/ignitetest/services/utils/ignite_aware.py
@@ -136,3 +136,14 @@ class IgniteAwareService(BackgroundThreadService, IgnitePersistenceAware, metacl
         for node in self.nodes:
             self.await_event_on_node(evt_message, node, timeout_sec, from_the_beginning=from_the_beginning,
                                      backoff_sec=backoff_sec)
+
+    def ssh_output_all(self, cmd: str):
+        """
+        Execute ssh command on all nodes.
+        """
+        res = {}
+        for node in self.nodes:
+            data = node.account.ssh_output(cmd)
+            res[node.account.hostname] = data
+
+        return res
